requirements.txt
astroid==2.0.4
boto3==1.9.43
botocore==1.12.43
colorama==0.4.0
Django==1.11
django-filter==2.0.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.0
docutils==0.14
et-xmlfile==1.0.1
isort==4.3.4
jdcal==1.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
mongoengine==0.16.0
openpyxl==2.5.10
pylint==2.1.1
pymongo==3.7.2
python-dateutil==2.7.5
pytz==2018.7
s3transfer==0.1.13
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11